backend/foodgram/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.16
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.0.4
flake8==2.3.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
pep8==1.7.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pytest==7.2.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0
